src/m1/iedcs-server/requirements.txt
Django==1.7.7
dj-database-url==0.3.0
dj-static==0.0.6
django-appconf==0.6
django-compressor==1.4
djangorestframework==3.1.0
drf-nested-routers==0.9.0
python-geoip-geolite2
python-geoip
Crypto
